requirements.txt
streamlit==1.28.0
langchain==0.1.0
langchain-core==0.1.0
langchain-community==0.0.10
langchain-cohere==0.1.0
langchain-postgres==0.0.6
psycopg[binary]==3.1.18
networkx==3.2.1
python-dotenv==1.0.0
numpy==1.24.3
scikit-learn==1.3.2
pypdf==3.17.0
spacy==3.7.2
nltk==3.8.1
tqdm==4.66.1